Added database validation

+++ src/modules/autoRole/index.ts
@@ -1,5 +1,4 @@
 import { GuildMember } from "discord.js";
-import ms from 'ms'
 
 import BaseModule from "../../utils/structures/BaseModule";
 import { AutoRoleModel } from "../../database/schemas/AutoRole";
@@ -16,44 +15,44 @@ class AutoRoleModule extends BaseModule {
     }
 
     async give(member: GuildMember) {
-        const config = await AutoRoleModel.findOne({ guildId: member.guild.id });
+        const config = await AutoRoleModel.findOne({ guildId: member.guild.id }).catch(() => {});
         if(!config) return
         for(const role of config.roles) {
-            await member.roles.add(role.roleId).catch(err => console.log(err));
-            if(!role.time) return;
+            await member.roles.add(role.roleId).catch(() => {});
+            if(!role.time) return
 
             const memberAutoRole = await AutoRoleTimeModel.findOne({ guildId: member.guild.id, userId: member.id }).catch(() => {});
             if(memberAutoRole) {
                 if(memberAutoRole.roles.find(e => e.roleId === role.roleId)) return;
                 await memberAutoRole.updateOne({
                     $push: { roles: { roleId: role.roleId, time: Date.now() + role.time } },
-                }).catch(err => console.log(err));
+                }, { runValidators: true }).catch(() => {});
             } else {
                 await AutoRoleTimeModel.create({
                     guildId: member.guild.id,
                     userId: member.id,
                     roles: [{ roleId: role.roleId, time: Date.now() + role.time }],
-                }).catch(err => console.log(err));
+                }).catch(() => {});
             }
 
             setTimeout(async () => {
-                member.roles.remove(role.roleId).catch(err => console.log(err));
+                member.roles.remove(role.roleId).catch(() => {});
                 const memberAutoRole = await AutoRoleTimeModel.findOneAndUpdate({ guildId: member.guild.id, userId: member.id }, {
                     $pull: {
                         roles: { roleId: role.roleId }
                     }
-                }, { new: true });
+                }, { new: true, runValidators: true });
                 if(memberAutoRole && !memberAutoRole.roles.length) {
                     memberAutoRole.delete();
                 }
-
             }, role.time)
         };
     }
 }
 
 async function checkAutoRoles() {
-    const collections = await AutoRoleTimeModel.find();
+    const collections = await AutoRoleTimeModel.find().catch(() => {})
+    if(!collections) return
     for (const collection of collections) {
         const guild = await client.guilds.fetch(collection.guildId).catch(() => {})
         if(!guild) continue
@@ -66,10 +65,9 @@ async function checkAutoRoles() {
                     $pull: {
                         roles: { roleId: role.roleId }
                     }
-                }, { new: true }).catch(() => {});
-                if(coll && !coll.roles.length) {
-                    coll.delete().catch(() => {});
-                }
+                }, { new: true, runValidators: true }).catch(() => {});
+                
+                if(coll && !coll.roles.length) coll.delete().catch(() => {});
             } else {
                 setTimeout(async () => {
                     member.roles.remove(role.roleId).catch(() => {});
@@ -77,10 +75,9 @@ async function checkAutoRoles() {
                         $pull: {
                             roles: { roleId: role.roleId }
                         }
-                    }, { new: true }).catch(() => {});
-                    if(coll && !coll.roles.length) {
-                        coll.delete().catch(() => {});
-                    }
+                    }, { new: true, runValidators: true }).catch(() => {});
+
+                    if(coll && !coll.roles.length) coll.delete().catch(() => {});
                 }, role.time - Date.now())
             }
         }
